requirements.txt
attrs==21.2.0
Brian2==2.5.0.1
brian2tools==0.3
cycler==0.11.0
Cython==0.29.25
fonttools==4.28.3
iniconfig==1.1.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
libNeuroML==0.2.58
lxml==4.7.1
markdown-strings==3.3.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mne==0.24.1
mpmath==1.2.1
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
PyLEMS==0.5.8
pyparsing==3.0.6
pyspikes==0.2.9
pytest==6.2.5
python-dateutil==2.8.2
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sympy==1.9
threadpoolctl==3.0.0
toml==0.10.2
tomli==2.0.0
